test setup from A2 integrated. Deadlock present (most likely due to OrleansClientManager setup)

+++ Infra/EventSchema/Checkout.cs
@@ -5,11 +5,11 @@ namespace Infra.EventSchema
     [MessagePackObject(keyAsPropertyName: true)]
     public class Checkout
     {
-        public int productId { get; set; }
+        public long productId { get; set; }
         public double price { get; set; }
         public int quantity { get; set; }
 
-        public Checkout(int productId, double price, int quantity)
+        public Checkout(long productId, double price, int quantity)
         {
             this.productId = productId;
             this.price = price;
